refactor: modify command from short_name to alias
